examples/wgpu: Set up for use on Android
