+++ examples/runners/wgpu/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2018"
 license = "MIT OR Apache-2.0"
 publish = false
 
+[lib]
+crate-type = ["lib", "cdylib"]
+
 # See rustc_codegen_spirv/Cargo.toml for details on these features
 [features]
 default = ["use-compiled-tools"]
@@ -19,6 +22,9 @@ winit = { version = "0.23", features = ["web-sys"] }
 
 [build-dependencies]
 spirv-builder = { path = "../../../crates/spirv-builder", default-features = false }
+
+[target.'cfg(target_os = "android")'.dependencies]
+ndk-glue = "0.2"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 wgpu-subscriber = "0.1.0"
